Reverted HttpClient back to original (removing virtual). Adjusted tests accordingly
