+++ src/ClientRuntime/Microsoft.Rest.ClientRuntime.Tests/Microsoft.Rest.ClientRuntime.Tests.xproj
@@ -17,5 +17,8 @@
   <ItemGroup>
     <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
   </ItemGroup>
+  <ItemGroup>
+    <DnxInvisibleCompile Include="CustomClients\FabrikamServiceClient.cs" />
+  </ItemGroup>
   <Import Project="$(VSToolsPath)\DotNet\Microsoft.DotNet.targets" Condition="'$(VSToolsPath)' != ''" />
 </Project>
